Seperate rule context out into new module
